requirements.txt
numpy
cmake
face_recognition
Pillow
matplotlib
streamlit
python-slugify